chore: Move spec files to the test directory
